Remving confusion between unsigned and signed types.

+++ zero/array.c
@@ -15,7 +15,7 @@ array_free(const struct gkyl_ref_count *ref)
 }
 
 struct gkyl_array*
-gkyl_array_new(size_t elemSz, size_t size)
+gkyl_array_new(long elemSz, long size)
 {
   struct gkyl_array* arr = gkyl_malloc(sizeof(struct gkyl_array));
   arr->elemSz = elemSz;
@@ -30,7 +30,7 @@ gkyl_array_copy(struct gkyl_array* dest, const struct gkyl_array* src)
 {
   assert(dest->elemSz == src->elemSz);
   
-  size_t ncopy = src->size < dest->size ? src->size : dest->size;
+  long ncopy = src->size < dest->size ? src->size : dest->size;
   memcpy(dest->data, src->data, ncopy*src->elemSz);
   return dest;
 }
